requirements.txt
discord.py
python-dotenv
langchain
langchain-groq
langchain-community
PyGithub
pyodbc
pytest
pytest-mock